requirements.txt
click==7.1.1
croniter==0.3.31
Flask==1.1.2
Flask-RQ2==18.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
redis==3.4.1
rq==1.3.0
rq-scheduler==0.9.1
six==1.14.0
Werkzeug==1.0.1
python-dotenv=0.14.0
requests
gunicorn
